+++ TravelersNotebook/Buttons.lua
@@ -1,15 +1,70 @@
 tb = tb or {}
 
+function tb.CreationInsertMenu()
+
+    tb.insertMenu = CreateFrame("Frame", nil, tb.frame, "BackdropTemplate")
+    tb.insertMenu:SetSize(120, 150)
+
+    tb.insertMenu:SetBackdrop({
+        bgFile = "Interface/Tooltips/UI-Tooltip-Background",
+        edgeFile = "Interface/Tooltips/UI-Tooltip-Border",
+        edgeSize = 12,
+    })
+
+    tb.insertMenu:SetBackdropColor(0, 0, 0, 0.9)
+    tb.insertMenu:SetFrameStrata("DIALOG")
+    tb.insertMenu:SetPoint("TOPLEFT", tb.menuButton, "BOTTOMLEFT", 0, -5)
+    tb.insertMenu:Hide()
+
+
+    tb.gpsButton = CreateFrame("Button", nil, tb.insertMenu, "UIPanelButtonTemplate")
+    tb.gpsButton:SetSize(75, 25)
+    tb.gpsButton:SetText("GPS")
+    tb.gpsButton:SetPoint("TOP", tb.insertMenu, "TOP", 0, -10)
+    tb.gpsButton:SetScript("OnClick", tb.CoordGPS)
+
+end
+
 -------------------------
 -- Génération des boutons
 -------------------------
 
 function tb.CreationBoutons()
 
+    -- bouton pour le sousmenu
+
+    tb.menuButton = CreateFrame("Button", nil, tb.frame, "UIPanelButtonTemplate")
+    tb.menuButton:SetSize(75, 25)
+    tb.menuButton:SetText("Insert")
+    tb.menuButton:SetScript("OnClick", function(self)
+        print(tb.insertMenu:GetWidth(), tb.insertMenu:GetHeight())
+        print(tb.insertMenu:GetLeft(), tb.insertMenu:GetTop())
+        if tb.insertMenu:IsShown() then
+            print("menu caché")
+            tb.insertMenu:Hide()
+        else
+            print("menu affiché")
+            tb.insertMenu:Show()
+        end
+    end)
+
+    tb.menuButton:SetScript("OnEnter", function(self)
+        GameTooltip:SetOwner(self, "ANCHOR_RIGHT")
+        GameTooltip:SetText("Afficher le sous menu")
+        GameTooltip:Show()
+    end)
+
+    tb.menuButton:SetScript("OnLeave", function()
+        GameTooltip:Hide()
+    end)
+
+    -- créer le contenu du menu
+    tb.CreationInsertMenu()
+
     -- bouton new note
 
     tb.newButton = CreateFrame("Button", nil, tb.frame, "UIPanelButtonTemplate")
-    tb.newButton:SetSize(85, 25)
+    tb.newButton:SetSize(75, 25)
     tb.newButton:SetText(tb.text.BUTTON_NEW)
     tb.newButton:SetScript("OnClick", tb.tnNewNote)
 
@@ -26,13 +81,13 @@ function tb.CreationBoutons()
     -- bouton save
 
     tb.saveButton = CreateFrame("Button", nil, tb.frame, "UIPanelButtonTemplate")
-    tb.saveButton:SetSize(85, 25)
+    tb.saveButton:SetSize(75, 25)
     tb.saveButton:SetText(tb.text.BUTTON_SAVE)
     tb.saveButton:SetScript("OnClick", tb.tnSaveNote)
 
     tb.saveButton:SetScript("OnEnter", function(self)
         GameTooltip:SetOwner(self, "ANCHOR_RIGHT")
-        GameTooltip:SetText("Save this note")
+        GameTooltip:SetText("Save current note")
         GameTooltip:Show()
     end)
 
@@ -43,7 +98,7 @@ function tb.CreationBoutons()
     -- bouton delete
 
     tb.deleteButton = CreateFrame("Button", nil, tb.frame, "UIPanelButtonTemplate")
-    tb.deleteButton:SetSize(85, 25)
+    tb.deleteButton:SetSize(75, 25)
     tb.deleteButton:SetText(tb.text.BUTTON_DEL)
     tb.deleteButton:SetScript("OnClick", tb.tnDeleteNote)
 
@@ -60,7 +115,7 @@ function tb.CreationBoutons()
     -- bouton about
 
     tb.aboutButton = CreateFrame("Button", nil, tb.frame, "UIPanelButtonTemplate")
-    tb.aboutButton:SetSize(85, 25)
+    tb.aboutButton:SetSize(75, 25)
     tb.aboutButton:SetText(tb.text.BUTTON_ABOUT)
     tb.aboutButton:SetScript("OnClick", tb.tnAboutWindows)
 
@@ -77,7 +132,7 @@ function tb.CreationBoutons()
     -- bouton Hide
 
     tb.hideButton = CreateFrame("Button", nil, tb.frame, "UIPanelButtonTemplate")
-    tb.hideButton:SetSize(85, 25)
+    tb.hideButton:SetSize(75, 25)
     tb.hideButton:SetText("Hide")
     tb.hideButton:SetScript("OnClick", tb.tnHideWindows)
 
@@ -91,27 +146,10 @@ function tb.CreationBoutons()
         GameTooltip:Hide()
     end)
 
-    -- bouton GPS
-
-    tb.gpsButton = CreateFrame("Button", nil, tb.frame, "UIPanelButtonTemplate")
-    tb.gpsButton:SetSize(85, 25)
-    tb.gpsButton:SetText("GPS")
-    tb.gpsButton:SetScript("OnClick", tb.CoordGPS)
-
-    tb.gpsButton:SetScript("OnEnter", function(self)
-        GameTooltip:SetOwner(self, "ANCHOR_RIGHT")
-        GameTooltip:SetText("Add your current location to the note")
-        GameTooltip:Show()
-    end)
-
-    tb.gpsButton:SetScript("OnLeave", function()
-        GameTooltip:Hide()
-    end)
-
-    -- bouton pour épingler
+    -- bouton pour fermer
 
     tb.closeNoteButton = CreateFrame("Button", nil, tb.frame, "UIPanelButtonTemplate")
-    tb.closeNoteButton:SetSize(85, 25)
+    tb.closeNoteButton:SetSize(75, 25)
     tb.closeNoteButton:SetText("Return")
     tb.closeNoteButton:SetScript("OnClick", tb.closeNote)
 
@@ -128,7 +166,7 @@ function tb.CreationBoutons()
     -- bouton pour épingler
 
     tb.pinButton = CreateFrame("Button", nil, tb.frame, "UIPanelButtonTemplate")
-    tb.pinButton:SetSize(85, 25)
+    tb.pinButton:SetSize(75, 25)
     tb.pinButton:SetText("Pin")
     tb.pinButton:SetScript("OnClick", tb.pinNote)
 
@@ -148,9 +186,9 @@ function tb.CreationBoutons()
         tb.newButton,
         tb.saveButton,
         tb.deleteButton,
-        tb.gpsButton,
         tb.closeNoteButton,
         tb.pinButton,
+        tb.menuButton,
         tb.aboutButton,
         tb.hideButton
     }
@@ -167,8 +205,8 @@ function tb.CreationBoutons()
         EDITION = {
             tb.saveButton,
             tb.deleteButton,
-            tb.gpsButton,
             tb.pinButton,
+            tb.menuButton,
             tb.closeNoteButton,
             tb.hideButton
         }